# lib/api-zod/src/generated/types/battlecard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Objection } from './objection';

export interface Battlecard {
  id: number;
  competitorId: number;
  competitorName: string;
  ourStrengths: string[];
  theirStrengths: string[];
  objections: Objection[];
  winThemes: string[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}