# lib/api-zod/src/generated/types/battlecardUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Objection } from './objection';

export interface BattlecardUpdate {
  ourStrengths?: string[];
  theirStrengths?: string[];
  objections?: Objection[];
  winThemes?: string[];
  notes?: string;
}