# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitorTier } from './competitorTier';

export interface Competitor {
  id: number;
  name: string;
  website: string;
  tier: CompetitorTier;
  /** @nullable */
  tagline?: string | null;
  strengths: string[];
  weaknesses: string[];
  targetSegment: string;
  /** @nullable */
  pricing?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}