# lib/api-zod/src/generated/types/competitorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitorInputTier } from './competitorInputTier';

export interface CompetitorInput {
  name: string;
  website: string;
  tier: CompetitorInputTier;
  tagline?: string;
  strengths: string[];
  weaknesses: string[];
  targetSegment: string;
  pricing?: string;
  notes?: string;
}