# lib/api-zod/src/generated/types/competitorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitorUpdateTier } from './competitorUpdateTier';

export interface CompetitorUpdate {
  name?: string;
  website?: string;
  tier?: CompetitorUpdateTier;
  tagline?: string;
  strengths?: string[];
  weaknesses?: string[];
  targetSegment?: string;
  pricing?: string;
  notes?: string;
}