# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Competitor } from './competitor';
import type { Signal } from './signal';

export interface DashboardSummary {
  competitorCount: number;
  icpCount: number;
  battlecardCount: number;
  signalCount: number;
  recentSignals: Signal[];
  topCompetitors: Competitor[];
}