# lib/api-zod/src/generated/types/icpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface IcpInput {
  name: string;
  industry: string;
  companySize: string;
  jobTitles?: string[];
  painPoints: string[];
  goals: string[];
  channels: string[];
  notes?: string;
}