# lib/api-zod/src/generated/types/objection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GTM Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Objection {
  objection: string;
  response: string;
}